+++ lib/src/models/mhtml_save_options_data.dart
@@ -377,6 +377,12 @@ class MhtmlSaveOptionsData extends HtmlSaveOptionsData {
       prettyFormat = null;
     }
 
+    if (json.containsKey('ReplaceBackslashWithYenSign')) {
+      replaceBackslashWithYenSign = json['ReplaceBackslashWithYenSign'] as bool;
+    } else {
+      replaceBackslashWithYenSign = null;
+    }
+
     if (json.containsKey('ResolveFontNames')) {
       resolveFontNames = json['ResolveFontNames'] as bool;
     } else {